% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.R
\name{gdm_fwdsel}
\alias{gdm_fwdsel}
\title{Perform forwards selection on GDM model context}
\usage{
gdm_fwdsel(sitepair, explained.tolerance = 0.1, max.vars = NULL)
}
\arguments{
\item{samples}{Sample names}

\item{respDist}{Response distance}

\item{longLatPreds}{Predictors. Cols must be Longitude, Latitude then other preds}

\item{minGeoDist}{Minimum geographic distance to allow between sites (filters site-pair table), in kilometers}

\item{nperm}{Number of permutations}

\item{sampleFrac}{Train on data at `sampleFrac*nSites` sites}
}
\value{
Data frame of per-predictor splines
}
\description{
Perform forwards selection on GDM model context
}
