% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.R
\name{gdm_jackknife_filtered}
\alias{gdm_jackknife_filtered}
\title{Runs a site-level jack-knifed GDM analysis}
\usage{
gdm_jackknife_filtered(samples, respDist, longLatPreds, minGeoDist = 1,
  nperm = 100, sampleFrac = 0.8, splines = NULL, knots = NULL)
}
\arguments{
\item{samples}{Sample names}

\item{respDist}{Response distance}

\item{longLatPreds}{Predictors. Cols must be Longitude, Latitude then other preds}

\item{minGeoDist}{Minimum geographic distance to allow between sites (filters site-pair table), in kilometers}

\item{nperm}{Number of permutations}

\item{sampleFrac}{Keep data at `sampleFrac*nSites` sites}
}
\value{
Data frame of per-predictor splines
}
\description{
Runs a site-level jack-knifed GDM analysis
}
