% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.R
\name{gower_cov2dist}
\alias{gower_cov2dist}
\title{Perform Gower's covariance to euclidean distance transform}
\usage{
gower_cov2dist(C)
}
\arguments{
\item{C}{Square, symmetrical covariance matrix}
}
\value{
D, normalised euclidean distance
}
\description{
Perform Gower's covariance to euclidean distance transform
}
