% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.R
\name{tidysplines}
\alias{tidysplines}
\title{Extract splines from a GDM model object into a tidy dataframe}
\usage{
tidysplines(gdm.mdl)
}
\arguments{
\item{gdm.mdl}{GDM model object}
}
\value{
A tidy data frame
}
\description{
Extract splines from a GDM model object into a tidy dataframe
}
